FEAT : community

See merge request s08-final/S08P31B204!230

+++ fri/src/main/java/com/project/fri/likes/repository/LikesRepository.java
@@ -11,5 +11,5 @@ import java.util.Optional;
 public interface LikesRepository extends JpaRepository<Likes, Long> {
 
     Optional<Likes> findByUserAndBoardAndIsDeleteFalse(User user, Board board);
-    int countByBoardAndIsDeleteFalse(Board board);
+    long countByBoardAndIsDeleteFalse(Board board);
 }
